Clean the code duplication
