+++ src/argouml-app/src/org/argouml/uml/ui/foundation/core/ActionAddSupplierDependencyAction.java
@@ -41,10 +41,8 @@ package org.argouml.uml.ui.foundation.core;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
-import org.argouml.i18n.Translator;
 import org.argouml.kernel.ProjectManager;
 import org.argouml.model.Model;
 import org.argouml.uml.ui.AbstractActionAddModelElement2;
@@ -91,42 +89,6 @@ public class ActionAddSupplierDependencyAction extends
         }
         ProjectManager.getManager().getCurrentProject()
                 .moveToTrash(toBeDeleted);
-    }
-
-    /*
-     * @see org.argouml.uml.ui.AbstractActionAddModelElement#getChoices()
-     */
-    protected List getChoices() {
-        List ret = new ArrayList();
-        Object model =
-            ProjectManager.getManager().getCurrentProject().getModel();
-        if (getTarget() != null) {
-            Object modelElementType = Model.getMetaTypes().getModelElement();
-            
-            ret.addAll(Model.getModelManagementHelper()
-                    .getAllModelElementsOfKind(model, modelElementType));
-            ret.remove(getTarget());
-        }
-        return ret;
-    }
-
-    /*
-     * @see org.argouml.uml.ui.AbstractActionAddModelElement#getDialogTitle()
-     */
-    protected String getDialogTitle() {
-        return Translator.localize("dialog.title.add-supplier-dependency");
-    }
-
-    /*
-     * @see org.argouml.uml.ui.AbstractActionAddModelElement#getSelected()
-     */
-    protected List getSelected() {
-        List v = new ArrayList();
-        Collection c =  Model.getFacade().getSupplierDependencies(getTarget());
-        for (Object supplierDependency : c) {
-            v.addAll(Model.getFacade().getClients(supplierDependency));
-        }
-        return v;
     }
 
 }
